refactor(ai-tools): move resolveCompanyToProjectIdFilter to filter-builder
